+++ mass_profile/fit_direct_beta.cpp
@@ -22,7 +22,7 @@ int main(int argc,char* argv[])
     }
 
   fitter<double,double,vector<double>,double,string> f;
-  
+
   f.set_statistic(chisq<double,double,vector<double>,double,string>());
   f.set_opt_method(powell_method<double,vector<double> >());
   f.set_model(beta1d<double>());
@@ -31,11 +31,11 @@ int main(int argc,char* argv[])
   ifs>>dl;
   f.load_data(dl.get_data_set());
   f.fit();
-  
+
   double rmin=f.get_data_set().get_data(0).get_x();
   double rmax=f.get_data_set().get_data(f.get_data_set().size()-1).get_x();
   cout<<"read terr 1 2\nskip single\n";
-  for(int i=0;i<f.get_data_set().size();++i)
+  for(size_t i=0;i<f.get_data_set().size();++i)
     {
       cout<<f.get_data_set().get_data(i).get_x()<<"\t"<<
 	-abs(f.get_data_set().get_data(i).get_x_lower_err())<<"\t"<<
@@ -44,7 +44,7 @@ int main(int argc,char* argv[])
 	-abs(f.get_data_set().get_data(i).get_y_lower_err())<<"\t"<<
 	abs(f.get_data_set().get_data(i).get_y_upper_err())<<endl;
 
-	
+
     }
   cout<<"no no no\n";
 
@@ -53,7 +53,7 @@ int main(int argc,char* argv[])
       cout<<i<<"\t0\t0\t"<<f.eval_model(i,f.get_all_params())<<"\t0\t0"<<endl;
     }
 
-  for(int i=0;i<f.get_num_params();++i)
+  for(size_t i=0;i<f.get_num_params();++i)
     {
       cerr<<f.get_param_info(i).get_name()<<"="<<
 	f.get_param_info(i).get_value()<<endl;
